Avoid errors when recreating an existing symlink on msys2
In this environment, symlinks don't exist but are made as copies.
In such cases, the -n option to ln doesn't work for making this
a no-op.

Alternatively, we could append "|| true", which would print an
error but continue.

+++ build-mingw-w64.sh
@@ -76,7 +76,9 @@ if [ -z "$HOST" ]; then
     if [ -z "$SKIP_INCLUDE_TRIPLET_PREFIX" ]; then
         for arch in $ARCHS; do
             mkdir -p $PREFIX/$arch-w64-mingw32
-            ln -sfn ../generic-w64-mingw32/include $PREFIX/$arch-w64-mingw32/include
+            if [ ! -e $PREFIX/$arch-w64-mingw32/include ]; then
+                ln -sfn ../generic-w64-mingw32/include $PREFIX/$arch-w64-mingw32/include
+            fi
         done
     fi
